requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.7.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
